pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "specpulse"
dynamic = ["version"]
description = "SpecPulse v2.2.4 - AI-Enhanced Specification-Driven Development Framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SpecPulse"}
]
maintainers = [
    {name = "SpecPulse"}
]
keywords = [
    "development",
    "specification",
    "ai",
    "claude",
    "gemini",
    "automation",
    "framework",
    "cli"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Documentation",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

dependencies = [
    "pyyaml>=6.0,<7.0",
    "click>=8.0,<9.0",
    "rich>=13.0,<14.0",
    "jinja2>=3.0,<4.0",
    "gitpython>=3.1,<4.0",
    "toml>=0.10,<1.0",
    "packaging>=21.0,<25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[project.scripts]
specpulse = "specpulse.cli.main:main"

[project.urls]
Homepage = "https://github.com/specpulse"
Repository = "https://github.com/specpulse/specpulse"
Issues = "https://github.com/specpulse/specpulse/issues"
Documentation = "https://github.com/specpulse/specpulse/wiki"
Changelog = "https://github.com/specpulse/specpulse/releases"

[tool.setuptools.dynamic]
version = {attr = "specpulse._version.__version__"}

[tool.setuptools]
packages = [
    "specpulse",
    "specpulse.cli",
    "specpulse.cli.commands",
    "specpulse.cli.handlers",
    "specpulse.cli.parsers",
    "specpulse.core",
    "specpulse.core.validators",
    "specpulse.utils",
    "specpulse.models",
    "specpulse.resources",
    "specpulse.resources.commands",
    "specpulse.resources.commands.claude",
    "specpulse.resources.commands.gemini",
    "specpulse.resources.memory",
]

[tool.setuptools.package-data]
specpulse = [
    "resources/memory/*.md",
    "resources/commands/claude/*.md",
    "resources/commands/claude/**/*.md",
    "resources/commands/gemini/*.toml",
    "resources/commands/gemini/**/*.toml",
    "resources/validation_rules.yaml",
    "resources/validation_examples.yaml",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]

[tool.coverage.run]
source = ["specpulse"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]